backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
azure-core==1.31.0
azure-identity==1.19.0
azure-monitor-opentelemetry-exporter==1.0.0b31
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.1
debugpy==1.8.7
decouple==0.0.7
Deprecated==1.2.14
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.2
fixedint==0.1.6
h11==0.14.0
idna==3.10
importlib_metadata==8.4.0
isodate==0.7.2
msal==1.31.0
msal-extensions==1.2.0
msrest==0.7.1
oauthlib==3.2.2
opentelemetry-api==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
portalocker==2.10.1
psutil==5.9.8
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pyodbc==5.1.0
python-decouple==3.8
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.40.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
wrapt==1.16.0
zipp==3.20.2
